+++ platform-sdk/swirlds-logging/src/test/java/com/swirlds/logging/ostream/BufferedOutputStreamTest.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package com.swirlds.logging.buffer;
+package com.swirlds.logging.ostream;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.junit.jupiter.api.Assertions.assertThrows;
